x86/ftrace: Adapt ftrace nop patching for PIE support
When using PIE with function tracing, the compiler generates a call
through the GOT (call *__fentry__@GOTPCREL). This instruction takes
6-bytes instead of 5-bytes with a relative call. And -mnop-mcount option
is not implemented for -fPIE now.

If PIE is enabled, replace the 6th byte of the GOT call by a 5-byte nop
and a 1-byte nop, so ftrace can handle the previous 5-bytes as before.

[Hou Wenlong: Adapt to code change and fix wrong offset calculation in
make_nop_x86()]

Signed-off-by: Thomas Garnier <thgarnie@chromium.org>
Co-developed-by: Hou Wenlong <houwenlong.hwl@antgroup.com>
Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>