objtool: Adapt indirect call of __fentry__() for PIE support
When using PIE with function tracing, the compiler generates a call
through the Global Offset Table (GOT) (call *fentry@GOTPCREL). This
instruction is an indirect call (INSN_CALL_DYNAMIC) and won't be
collected by add_call_destinations().  To patch it as NOP, these
indirect calls of __fentry__() must be collected individually.
Additionally, it needs to replace the 6-byte GOT indirect call with a
5-byte NOP followed by a 1-byte NOP so that ftrace can handle the
previous 5-byte instruction as before.

When RETPOLINE is enabled, __fentry__() is still an indirect call, which
generates warnings in objtool. However, it would be patched as NOP
later, for simplicity, regard it as INSN_CALL to omit warnings for jump
table and retpoline checks in ojbtool.

Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>