x86/stackprotector/64: Use per-cpu stack canary if supported by compiler
GCC 8.1 added options that allow the stack canary to be accessed as
"%gs:__stack_chk_guard", effectively turning it into an ordinary per-cpu
variable. For CONFIG_SMP=n, it is not necessary to set up GSBASE as it
could be set as 0. So use the per-cpu stack canary if supported by the
compiler and keep the fixed location code for compatibility with older
compilers.

Co-developed-by: Brian Gerst <brgerst@gmail.com>
Signed-off-by: Brian Gerst <brgerst@gmail.com>
Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>