x86/head/64: Build the head code as PIE
During the early boot stage, the head code runs in a low identity
address, so all absolute references would be incorrect. However, the
compiler doesn't have to generate PC-relative references when accessing
globals. To work around this problem, every global variable access must
be adjusted using fixup_pointer(). Nevertheless, the compiler could
generate absolute references for some global variable accesses, and the
behavior differs between GCC and CLANG. For example, GCC generates
PC-relative references for 'next_early_pgt', while CLANG generates
absolute references. Moreover, the rule is not always clear, e.g.,
'pgdir_shift' is a non-static global variable similar to
'next_early_pgt', but the compiler could generate the right PC-relative
reference, so no fixup_pointer() is needed when using the PGDIR_SHIFT
macro.

To avoid such cases, build the head code as PIE to force the generation
of PC-relative references and eliminate the need for fixup_pointer().
However, it may need an absolute virtual address for some symbols in the
head code. Therefore, use 'movabsq' to obtain the absolute virtual
address if needed. However, the mcmodel=kernel option is not compatible
with the fPIE option, so it needs to be removed. Nonetheless, this would
result in using the '%fs' register for stack canary access if the stack
protector is enabled. Additionally, per-cpu stack protector is not
currently supported for 64-bit. Since the head code is in the early
booting stage, it is safe to disable the stack protector for the head
code.

Suggested-by: Lai Jiangshan <jiangshan.ljs@antgroup.com>
Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>