x86/pie: Allow kernel image to be relocated at any address
For PIE kernel image, it could be relocated at any address. To be
simplified, treat the 2G area which including kernel image, modules area
and fixmap area as a whole, allow it to be relocated at any kernel
address.  After that, the relocated kernel address may be below than
__START_KERNEL_map, so use a global variable to store the base of
relocated kernel image.  And pa/va transformation of kernel image
address is adapted.

Suggested-by: Lai Jiangshan <jiangshan.ljs@antgroup.com>
Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>