KVM: x86/PVM: Store the valid value for MSR_PVM_VCPU_STRUCT unconditionally
During the VM restore process, if the VMM (e.g., Cloud Hypervisor)
restores MSRs before adding the user memory region, it can result in a
failure in kvm_gpc_activate() because no memslot has been added yet. As
a consequence, the VM will panic after the VM restore since the GPC is
not active. However, if we store the value even if kvm_gpc_activate()
fails later when the GPC is active, it can be refreshed by the addition
of the user memory region before the VM entry.

Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>
Link: https://github.com/virt-pvm/linux/issues/2