KVM: x86/PVM: Keep MSR_STAR always associated with CS selector
According to the PVM specification, the CS selector pushed on the stack
during kernel event injection is obtained from the MSR_STAR. If the CS
selector is changed in CPL0, the MSR_STAR should also be updated
accordingly to ensure that the guest kernel receives the correct CS
selector during kernel event handling. Therefore, it is important to
always keep MSR_STAR associated with the CS selector.

Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>
Link: https://github.com/virt-pvm/linux/issues/3