KVM: x86: Introduce KVM_HC_PV_MMU_RELEASE_PT hypercall
The hypercall is used by the guest to notify the hypervisor of page
table reclamation. This allows the hypervisor to release the shadowed SP
directly, rather than detecting a write flood. In the future, with the
introduction of an enlightened shadow MMU mode, also called PV MMU mode,
the memory of the guest page table will not be write-protected,
rendering the write flood ineffective. So, guests should use this
hypercall after enabling the PV MMU mode.

Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>