x86/mm: Reserve a range in the 5-level page tables for PVM guest
For a 4-level paging mode PVM guest, only the top 128TB is canonical.
When KASLR is enabled for 5-level page tables, this range overlaps with
the KASLR entropy range. Therefore, set the end address to -128TB to
reserve a range for the PVM guest. Regarding the KASAN area, the size is
sufficient.

Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>