KVM: x86/PVM: Disable direct switching when the GPC is inactive
During direct switching, the switcher directly uses 'pvcs_gpc.khva' to
access the PVCS structure without performing any checks. However, NULL
pointer access can be easily triggered by the guest if it first writes a
valid value to MSR_PVM_VCPU_STRUCT to activate the GPC, and then writes
a zero to deactivate the GPC and set the pvcs_gpc.khva to NULL.
Therefore, a new switch flag is introduced to disable direct switching
if the GPC is inactive.

Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>