KVM: x86/PVM: Don't expose MSR_IA32_TSX_CTRL to guest
If MSR_IA32_TSX_CTRL is exposed to the guest, PVM needs to switch the
MSR value in hardware during VM enter/exit. Therefore, don't expose it
to the guest for simplicity. Additionally, if the VMM clears the RTM
feature in the guest CPUID but the host enables it in hardware, PVM also
needs to disable it by setting TSX_CTRL_RTM_DISABLE into
MSR_IA32_TSX_CTRL if it is supported by hardware. Otherwise, the guest
could use the RTM feature even though the related CPUID is missing.
However, this behavior is acceptable in VMX without MSR_IA32_TSX_CTRL
support in hardware. Therefore, allowing it in PVM as if the
MSR_IA32_TSX_CTRL is not supported in hardware.

Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>