KVM: x86/PVM: Add pvm_trace.h header files to tools
Add the pvm_trace.h header files to the tools to provide the necessary
header files for tools that require PVM support, such as perf.

The change should not have any negative impact on the existing tools or
the kernel itself, as it simply makes the header files available for use
by the tools that need them.

Signed-off-by: Jinrong Liang <cloudliang@tencent.com>