x86/pvm: Don't profile PVH entry code
When CONFIG_FUNCTION_TRACER is enabled, there will be a call to
__fentry__ before performing relocation in pvm_update_pgtable().
However, GCC will generate a GOT reference for __fentry__ under PIE
building, causing it to try to access the compile-address of __fentry__,
leading to booting failure. Since the PVH entry code is only used in
early booting, the ftrace compile flags for the
arch/x86/platform/pvh/enlighten.c file could be removed.

Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>
Link: https://github.com/virt-pvm/linux/issues/11