KVM: x86/PVM: Handle #VE events
The #VE events that occur during the running of a PVM guest would exit
to PVM, so PVM needs to handle it. The handling needs to be split into
two steps. The first step is to obtain the #VE information in
pvm_vcpu_run_noinstr() after the VM exits, as early as possible before
any further operations that may trigger a nested #VE. Then, the cached
information can be used in handling #VE in handle_exit_exception().
And only EXIT_REASON_CPUID is allowed for now.

Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>