x86/ftrace: Adapt ftrace nop patching for PIE support
When using PIE with function tracing, GCC versions before 14 generate a
call through the GOT (call *__fentry__@GOTPCREL). This instruction takes
6 bytes instead of 5 bytes with a direct call.  Although it can be
relaxed by the linker as a 5-byte direct call followed by a 1-byte nop,
it cannot be relaxed for modules. And -mnop-mcount option is not
implemented for -fPIE now.

If PIE is enabled, replace the 6-byte GOT indirect call by a 5-byte nop
and a 1-byte nop, so that ftrace can handle the previous 5-bytes as
before.

Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>