KVM: PVM: Introduce ASID to manage host PCID for guest
ntroduce ASID to manage the host PCID for the guest, where an ASID is a
group of host PCIDs. Each vCPU will be assigned an ASID when it is
loaded onto a pCPU, and each guest PCID on this vCPU is directly mapped
to its host PCID by adding the ASID of the vCPU. This approach can
improve direct switching and reduce the complexity of host PCID
allocation for the guest. However, this conflicts with the global ASID
allocation in the upstream, which will be addressed later.

Signed-off-by: Lai Jiangshan <jiangshan.ljs@antgroup.com>
Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>
Link: https://github.com/virt-pvm/linux/issues/20