Revert "KVM: Drop unused @may_block param from gfn_to_pfn_cache_invalidate_start()"
This reverts commit eefb85b3f0310c2f4149c50cb9b13094ed1dde25.

The reason is to get back the usage of 'KVM_GUEST_USES_PFN' for the PFN
cache. PVM relies on 'KVM_GUEST_USES_PFN' usage of the PFN cache to share
PVCS pages between the guest and host during direct switching.
Therefore, we are temporarily reverting it; we will redesign the PVCS
page sharing implementation later.

Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>