Revert "KVM: pfncache: remove KVM_GUEST_USES_PFN usage"
This reverts commit a4bff3df51472f555ab8dea05a3d2faf4abbf199.

The reason is to get back the usage of 'KVM_GUEST_USES_PFN' for the PFN
cache. PVM relies on 'KVM_GUEST_USES_PFN' usage of the PFN cache to
share PVCS pages between the guest and host during direct switching.
Therefore, we are temporarily reverting it; we will redesign the PVCS
page sharing implementation later.

Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>